third_party/WebKit/LayoutTests/editing/inserting/return-key-with-selection-002.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div class="explanation">
<div class="scenario">
Tests: 
<br>
A scenario I thought of based on my experiences with this bug: 
<a href="rdar://problem/4045521">&lt;rdar://problem/4045521&gt;</a> Hitting return key with full line selected does not add blank line as it should
</div>
<div class="expected-results">
Expected Results:
<br>
Should see this content in the box below (note that the insertion point should be at the start of the third line, immediately preceding "baz"):
<div>foo</div><div><br></div><div>baz</div>
</div>
</div>

<div contenteditable id="root" style="word-wrap: break-word; -khtml-nbsp-mode: space; -khtml-line-break: after-white-space;">
<div id="sample" class="editing"><div>foo</div><div>bar</div><div id="baz">bbaz</div></div>
</div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    var sample = document.getElementById('sample');

    selection.collapse(sample, 0);
    selection.modify('move', 'forward', 'line');
    selection.modify('extend', 'forward', 'line');
    selection.modify('extend', 'forward', 'character');
    document.execCommand('insertParagraph');

    assert_equals(sample.innerHTML, '<div>foo</div><div><br></div><div>baz<br></div>');
    assert_true(selection.isCollapsed);
    assert_equals(selection.anchorNode, sample.childNodes[2].firstChild);
    assert_equals(selection.anchorOffset, 0);
});
</script>
